Integrated and simplified add voxel for cube and sphere
