+++ crates/bevy_voxel/src/editstate/add_normal.rs
@@ -2,6 +2,8 @@ use bevy::prelude::*;
 use voxels::data::voxel_octree::VoxelMode;
 use crate::{EditState, Preview, BevyVoxelResource, Center, Chunks, PreviewGraphics, ChunkData, ShapeState, MeshComponent};
 
+use super::{EditEvents, EditEvent};
+
 
 pub struct CustomPlugin;
 impl Plugin for CustomPlugin {
@@ -15,70 +17,44 @@ impl Plugin for CustomPlugin {
 
 fn add_voxel_cube(
   mouse: Res<Input<MouseButton>>,
-  mut bevy_voxel_res: ResMut<BevyVoxelResource>,
-
-  mut chunks: Query<(&Preview, &Center, &mut Chunks, &mut MeshComponent)>,
+  mut chunks: Query<&Preview>,
   shape_state: Res<State<ShapeState>>,
+  
+  mut edit_event_writer: EventWriter<EditEvents>
 ) {
   if !mouse.just_pressed(MouseButton::Left) ||
   shape_state.0 != ShapeState::Cube {
     return;
   }
 
-  for (preview, center, mut chunks, mut mesh_comp) in &mut chunks {
+  for preview in &mut chunks {
     if preview.pos.is_none() {
       continue;
     }
-
-    let p = preview.pos.unwrap();
-    let res = bevy_voxel_res.set_voxel_cube(p, preview);
-
-    let mut all_chunks = Vec::new();
-    for (key, chunk) in res.iter() {
-      all_chunks.push(chunk.clone());
-      chunks.data.insert(*key, chunk.clone());
-    }
-
-    let data = bevy_voxel_res.load_mesh_data(&all_chunks);
-    for (mesh_data, handle) in data.iter() {
-      
-      mesh_comp.data.insert(mesh_data.key.clone(), mesh_data.clone());
-      mesh_comp.added.push((mesh_data.clone(), *handle));
-    }
+    edit_event_writer.send(EditEvents {
+      event: EditEvent::AddCube
+    });
   }
 }
 
 fn add_voxel_sphere(
   mouse: Res<Input<MouseButton>>,
-  mut bevy_voxel_res: ResMut<BevyVoxelResource>,
-
-  mut chunks: Query<(&Preview, &Center, &mut Chunks, &mut MeshComponent)>,
+  mut chunks: Query<&Preview>,
   shape_state: Res<State<ShapeState>>,
+  mut edit_event_writer: EventWriter<EditEvents>
 ) {
   if !mouse.just_pressed(MouseButton::Left) ||
   shape_state.0 != ShapeState::Sphere {
     return;
   }
 
-  for (preview, center, mut chunks, mut mesh_comp) in &mut chunks {
+  for preview in &mut chunks {
     if preview.pos.is_none() {
       continue;
     }
-
-    let p = preview.pos.unwrap();
-    let res = bevy_voxel_res.set_voxel_sphere(p, preview);
-
-    let mut all_chunks = Vec::new();
-    for (key, chunk) in res.iter() {
-      all_chunks.push(chunk.clone());
-      chunks.data.insert(*key, chunk.clone());
-    }
-
-    let data = bevy_voxel_res.load_mesh_data(&all_chunks);
-    for (mesh_data, handle) in data.iter() {
-      mesh_comp.data.insert(mesh_data.key.clone(), mesh_data.clone());
-      mesh_comp.added.push((mesh_data.clone(), *handle));
-    }
+    edit_event_writer.send(EditEvents {
+      event: EditEvent::AddSphere
+    });
   }
 }
