+++ crates/bevy_voxel/src/editstate/add_snap.rs
@@ -3,6 +3,8 @@ use utils::RayUtils;
 use voxels::data::voxel_octree::VoxelMode;
 use crate::{EditState, Preview, BevyVoxelResource, Center, Chunks, PreviewGraphics, ChunkData, ShapeState};
 
+use super::{EditEvents, EditEvent};
+
 
 pub struct CustomPlugin;
 impl Plugin for CustomPlugin {
@@ -99,67 +101,44 @@ fn set_distance(
 
 fn add_voxel_cube(
   mouse: Res<Input<MouseButton>>,
-  mut bevy_voxel_res: ResMut<BevyVoxelResource>,
-
-  mut chunks: Query<(&Preview, &Center, &mut Chunks)>,
+  mut chunks: Query<&Preview>,
   shape_state: Res<State<ShapeState>>,
+  
+  mut edit_event_writer: EventWriter<EditEvents>
 ) {
   if !mouse.just_pressed(MouseButton::Left) ||
   shape_state.0 != ShapeState::Cube {
     return;
   }
 
-  for (preview, center, mut chunks) in &mut chunks {
+  for preview in &mut chunks {
     if preview.pos.is_none() {
       continue;
     }
-
-    chunks.data.clear();
-    let p = preview.pos.unwrap();
-    bevy_voxel_res.set_voxel_cube(p, preview);
-
-    let all_chunks = bevy_voxel_res.load_adj_chunks_with_collider(center.key);
-    for chunk in all_chunks.iter() {
-      let data = bevy_voxel_res.compute_mesh(VoxelMode::SurfaceNets, chunk);
-      if data.positions.len() == 0 {
-        continue;
-      }
-      
-      chunks.data.insert(chunk.key, chunk.clone());
-    }
+    edit_event_writer.send(EditEvents {
+      event: EditEvent::AddCube
+    });
   }
 }
 
 fn add_voxel_sphere(
   mouse: Res<Input<MouseButton>>,
-  mut bevy_voxel_res: ResMut<BevyVoxelResource>,
-
-  mut chunks: Query<(&Preview, &Center, &mut Chunks)>,
+  mut chunks: Query<&Preview>,
   shape_state: Res<State<ShapeState>>,
+  mut edit_event_writer: EventWriter<EditEvents>
 ) {
   if !mouse.just_pressed(MouseButton::Left) ||
   shape_state.0 != ShapeState::Sphere {
     return;
   }
 
-  for (preview, center, mut chunks) in &mut chunks {
+  for preview in &mut chunks {
     if preview.pos.is_none() {
       continue;
     }
-
-    chunks.data.clear();
-    let p = preview.pos.unwrap();
-    bevy_voxel_res.set_voxel_sphere(p, preview);
-
-    let all_chunks = bevy_voxel_res.load_adj_chunks_with_collider(center.key);
-    for chunk in all_chunks.iter() {
-      let data = bevy_voxel_res.compute_mesh(VoxelMode::SurfaceNets, chunk);
-      if data.positions.len() == 0 {
-        continue;
-      }
-      
-      chunks.data.insert(chunk.key, chunk.clone());
-    }
+    edit_event_writer.send(EditEvents {
+      event: EditEvent::AddSphere
+    });
   }
 }
